Fix month creation. #name# is really used within JabRef to determine a BibTeX string content (and not plain text)
- Fix tests
- Sort BibEntryTests: default constructor, keywords, all others
- Fix name of "CanonicalBibEntry" (from CanonicalBibtexEntry)
- Revert "fix jabref format of month", this also reverts commit 77c25ca0f1702991493cf37278a4cedad0d984a4.
- Format some code
- Add comments
