+++ src/main/java/org/jabref/model/entry/BibEntry.java
@@ -614,7 +614,7 @@ public class BibEntry implements Cloneable {
      */
     @Override
     public String toString() {
-        return CanonicalBibtexEntry.getCanonicalRepresentation(this);
+        return CanonicalBibEntry.getCanonicalRepresentation(this);
     }
 
     /**
