+++ src/main/java/org/jabref/model/entry/Month.java
@@ -168,13 +168,14 @@ public enum Month {
      * Returns the month in JabRef format. The format is the short 3-digit name surrounded by a '#'.
      * Example: #jan#, #feb#, etc.
      *
-     * See https://github.com/JabRef/jabref/issues/263#issuecomment-151246595 for a discussion on that thing.
-     * This seems to be an <em>invalid</em> format in terms of plain BiBTeX, but a <em>valid</em> format in the case of JabRef
+     * See <a href="https://github.com/JabRef/jabref/issues/263#issuecomment-151246595">Issue 263</a> for a discussion on that thing.
+     * This seems to be an <em>invalid</em> format in terms of plain BiBTeX, but a <em>valid</em> format in the case of JabRef.
+     * The documentation is available at the <a href="http://help.jabref.org/en/Strings">Strings help</a> of JabRef.
      *
      * @return Month in JabRef format
      */
     public String getJabRefFormat() {
-        return String.format("%s", shortName);
+        return String.format("#%s#", shortName);
     }
 
     /**
