+++ src/test/java/org/jabref/model/entry/BibEntryTest.java
@@ -1,141 +1,73 @@
 package org.jabref.model.entry;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
-import org.jabref.model.database.BibDatabase;
-import org.jabref.model.entry.field.BibField;
-import org.jabref.model.entry.field.FieldPriority;
+import org.jabref.model.entry.field.InternalField;
+import org.jabref.model.entry.field.OrFields;
 import org.jabref.model.entry.field.StandardField;
-import org.jabref.model.entry.field.UnknownField;
+import org.jabref.model.entry.types.StandardEntryType;
 
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class BibEntryTest {
+/**
+ * Contains misc tests for BibEntry especially not using the default constructor: {@link
+ * BibEntryWithDefaultConstructorTest}
+ */
+class BibEntryTest {
+    @Test
+    public void allFieldsPresentDefault() {
+        BibEntry e = new BibEntry(StandardEntryType.Article);
+        e.setField(StandardField.AUTHOR, "abc");
+        e.setField(StandardField.TITLE, "abc");
+        e.setField(StandardField.JOURNAL, "abc");
 
-    private BibEntry entry;
+        List<OrFields> requiredFields = new ArrayList<>();
+        requiredFields.add(new OrFields(StandardField.AUTHOR));
+        requiredFields.add(new OrFields(StandardField.TITLE));
+        assertTrue(e.allFieldsPresent(requiredFields, null));
 
-    @BeforeEach
-    public void setUp() {
-        entry = new BibEntry();
-    }
-
-    @AfterEach
-    public void tearDown() {
-        entry = null;
+        requiredFields.add(new OrFields(StandardField.YEAR));
+        assertFalse(e.allFieldsPresent(requiredFields, null));
     }
 
     @Test
-    public void getFieldIsCaseInsensitive() throws Exception {
-        entry.setField(new UnknownField("TeSt"), "value");
+    public void allFieldsPresentOr() {
+        BibEntry e = new BibEntry(StandardEntryType.Article);
+        e.setField(StandardField.AUTHOR, "abc");
+        e.setField(StandardField.TITLE, "abc");
+        e.setField(StandardField.JOURNAL, "abc");
 
-        assertEquals(Optional.of("value"), entry.getField(new UnknownField("tEsT")));
+        List<OrFields> requiredFields = new ArrayList<>();
+        requiredFields.add(new OrFields(StandardField.JOURNAL, StandardField.YEAR));
+        assertTrue(e.allFieldsPresent(requiredFields, null));
+
+        requiredFields.add(new OrFields(StandardField.YEAR, StandardField.ADDRESS));
+        assertFalse(e.allFieldsPresent(requiredFields, null));
     }
 
     @Test
-    public void getFieldWorksWithBibFieldAsWell() throws Exception {
-        entry.setField(StandardField.AUTHOR, "value");
-
-        assertEquals(Optional.of("value"), entry.getField(new BibField(StandardField.AUTHOR, FieldPriority.IMPORTANT).getField()));
+    public void isNullCiteKeyThrowsNPE() {
+        BibEntry e = new BibEntry(StandardEntryType.Article);
+        assertThrows(NullPointerException.class, () -> e.setCiteKey(null));
     }
 
     @Test
-    public void setFieldWorksWithBibFieldAsWell() throws Exception {
-        entry.setField(new BibField(StandardField.AUTHOR, FieldPriority.IMPORTANT).getField(), "value");
+    public void isEmptyCiteKey() {
+        BibEntry e = new BibEntry(StandardEntryType.Article);
+        assertFalse(e.hasCiteKey());
 
-        assertEquals(Optional.of("value"), entry.getField(StandardField.AUTHOR));
-    }
+        e.setCiteKey("");
+        assertFalse(e.hasCiteKey());
 
-    @Test
-    public void clonedBibentryHasUniqueID() throws Exception {
-        BibEntry entry = new BibEntry();
-        BibEntry entryClone = (BibEntry) entry.clone();
+        e.setCiteKey("key");
+        assertTrue(e.hasCiteKey());
 
-        assertNotEquals(entry.getId(), entryClone.getId());
-    }
-
-    @Test
-    public void testGetAndAddToLinkedFileList() {
-        List<LinkedFile> files = entry.getFiles();
-        files.add(new LinkedFile("", "", ""));
-        entry.setFiles(files);
-        assertEquals(Arrays.asList(new LinkedFile("", "", "")), entry.getFiles());
-    }
-
-    @Test
-    public void testGetEmptyKeywords() {
-        KeywordList actual = entry.getKeywords(',');
-
-        assertEquals(new KeywordList(), actual);
-    }
-
-    @Test
-    public void testGetSingleKeywords() {
-        entry.addKeyword("kw", ',');
-        KeywordList actual = entry.getKeywords(',');
-
-        assertEquals(new KeywordList(new Keyword("kw")), actual);
-    }
-
-    @Test
-    public void testGetKeywords() {
-        entry.addKeyword("kw", ',');
-        entry.addKeyword("kw2", ',');
-        entry.addKeyword("kw3", ',');
-        KeywordList actual = entry.getKeywords(',');
-
-        assertEquals(new KeywordList(new Keyword("kw"), new Keyword("kw2"), new Keyword("kw3")), actual);
-    }
-
-    @Test
-    public void testGetEmptyResolvedKeywords() {
-        BibDatabase database = new BibDatabase();
-        BibEntry entry2 = new BibEntry();
-        entry.setField(StandardField.CROSSREF, "entry2");
-        entry2.setCiteKey("entry2");
-        database.insertEntry(entry2);
-        database.insertEntry(entry);
-
-        KeywordList actual = entry.getResolvedKeywords(',', database);
-
-        assertEquals(new KeywordList(), actual);
-    }
-
-    @Test
-    public void testGetSingleResolvedKeywords() {
-        BibDatabase database = new BibDatabase();
-        BibEntry entry2 = new BibEntry();
-        entry.setField(StandardField.CROSSREF, "entry2");
-        entry2.setCiteKey("entry2");
-        entry2.addKeyword("kw", ',');
-        database.insertEntry(entry2);
-        database.insertEntry(entry);
-
-        KeywordList actual = entry.getResolvedKeywords(',', database);
-
-        assertEquals(new KeywordList(new Keyword("kw")), actual);
-    }
-
-    @Test
-    public void testGetResolvedKeywords() {
-        BibDatabase database = new BibDatabase();
-        BibEntry entry2 = new BibEntry();
-        entry.setField(StandardField.CROSSREF, "entry2");
-        entry2.setCiteKey("entry2");
-        entry2.addKeyword("kw", ',');
-        entry2.addKeyword("kw2", ',');
-        entry2.addKeyword("kw3", ',');
-        database.insertEntry(entry2);
-        database.insertEntry(entry);
-
-        KeywordList actual = entry.getResolvedKeywords(',', database);
-
-        assertEquals(new KeywordList(new Keyword("kw"), new Keyword("kw2"), new Keyword("kw3")), actual);
+        e.clearField(InternalField.KEY_FIELD);
+        assertFalse(e.hasCiteKey());
     }
 }
