+++ src/test/resources/testbib/saveactions2.bib
@@ -7,7 +7,7 @@
   year =      {2015},
   pages =     {1--20},
   address =   {Hamburg, Germany},
-  month =     mar
+  month =     mar,
   file =      {:Pautasso2015TowardsWorkflowBenchmarking.pdf:PDF},
   keywords =  {benchflow}
 }
